commit_dec_18_blueprint_admin
