+++ app/admin/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+bp = Blueprint('admin', __name__)
+
+
+from app.admin import forms, routes
