+++ app/templates/test.html
@@ -90,7 +90,7 @@
                     </div>
                     <div class="row">
                         <div class="file col-xl-9 col-lg-9 col-sm-9 col-xs-12">
-                                {% if i in (3,9) %}
+                                {% if i in (3,9,22) %}
 
                             <ul>
                                 <li><a href="https://kege.it-generation.ru/tasks/{{ test_path}}files/0{{i}}.xlsx" download>Скачать файл</a></li>
